+++ frontend/src/components/IssueCreation/CommentInput.js
@@ -1,59 +1,12 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import ReactMarkdown from 'react-markdown';
-import gfm from 'remark-gfm';
+import MarkdownEditor from '../Common/MarkdownEditor';
 
 export default ({ tabType, setContent, content }) => {
   const onChange = e => {
     setContent(e.target.value);
   };
 
-  const InlineCodeBlock = props => {
-    return <span style={{ fontWeight: 'bold' }}>{props.value}</span>;
-  };
-
-  const BlockQuoteBlock = props => {
-    return (
-      <div
-        style={{
-          border: '1px dashed #aaa',
-          borderRadius: 10,
-          paddingLeft: 10,
-          margin: 5,
-        }}
-      >
-        {props.children}
-      </div>
-    );
-  };
-
-  const CodeBlock = props => {
-    return (
-      <pre style={{ background: 'rgba(0,0,0,0.1)', padding: 10 }}>
-        <code>{props.value}</code>
-      </pre>
-    );
-  };
-
-  const TableCellBlock = props => {
-    let style = {
-      textAlign: props.align ? props.align : 'center',
-      padding: 5,
-    };
-
-    if (props.isHeader) {
-      style.background = 'rgba(0,0,0,0.1)';
-      style.border = '1ps solid #ccc';
-      style.borderLeft = 0;
-      style.borderRight = 0;
-    } else {
-      style.borderLeft = '1px solid #eee';
-      style.borderRight = '1px solid #eee';
-      style.borderBottom = '1px solid #eee';
-    }
-
-    return <td style={style}>{props.children}</td>;
-  };
   return (
     <>
       {tabType === 'write' && (
@@ -65,18 +18,7 @@ export default ({ tabType, setContent, content }) => {
       )}
       {tabType === 'preview' && (
         <MarkdownContent>
-          <ReactMarkdown
-            plugins={[gfm]}
-            source={content}
-            skipHtml={false}
-            allowDangerousHtml
-            renderers={{
-              code: CodeBlock,
-              tableCell: TableCellBlock,
-              inlineCode: InlineCodeBlock,
-              blockquote: BlockQuoteBlock,
-            }}
-          />
+          <MarkdownEditor content={content} />
         </MarkdownContent>
       )}
     </>
